dsc/uge07/matrix_image_processing/pyproject.toml
[tool.poetry]
name = "matrix_image_processing"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
nose2 = "^0.10.0"
click = "^7.1.2"
Pillow = "^8.1.2"
numpy = "^1.20.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
